requirements.txt
langgraph>=0.0.15
langchain>=0.1.0
langchain-anthropic>=0.1.1
pydantic>=2.5.2
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
google-cloud-aiplatform>=1.36.4
amadeus>=9.0.0
googlemaps>=4.10.0 